V1.3
When reading the program for the first time, the program will get the words from the original file. A copy of this file is created and every time the user knows a word and presses the button, that word is removed from the 'coppied' list.

Doing so, every time the user runs the program, it will only apear words he dosen't know the meaning